feat(VDP): Refactor `VDP2ComposeLine` into separable passes
This refactors `VDP2ComposeLine` into multiple passes of "kernels" to better lend itself to SIMD optimizations and to opportunistically detect uniform control flow.

Conditionals are computed upfront into lists of `bool`s that can be checked quickly for uniformity(`all`/`any`/`none`). Entire passes can then be avoided depending on how uniform(or not) the conditional data is.

Ex, Doing advanced blending vs a simple memcpy if it is found that the entire scanline does not utilize blending.