build: Use Clang 21 in FreeBSD CI and switch to latest repository
Arm64 builds require at least Clang 21 due to a compiler bug in lower versions.
While arm64 builds take a long time in the CI and are thus not enabled, building
at least for amd64 with Clang 21 helps to ensure nothing breaks.

Enable the latest package repository to have llvm21 available in CI. This can be
reverted when llvm21 hits the 2025Q4 quarterly repository.