build: Fix FreeBSD CI with vcpkg
FreeBSD's CI is restructured to match the other platforms.

The CXX and CC environment variables are set and propagated to vcpkg and
CMake in order for them to pick up the same compiler. The path to vcpkg's
sdl3.pc is added to the PKG_CONFIG_PATH environment variable because
otherwise vcpkg is unable to locate it for some unknown reason.

The following new packages have to be installed for vcpkg to build the
packages successfully: git, gmake, patchelf, python3 and zip.

With the usage of vcpkg it's now not necessary anymore to propagate the
/usr/local/include und /usr/local/lib paths to the build system.

COMPILING.md has been updated accordingly. It also is a little bit more
streamlined.

The vcpkg cache is currently not used. This needs more investigation and
testing.